Reuse container preparation code
