+++ src/ol/renderer/canvas/ImageLayer.js
@@ -7,8 +7,6 @@ import ViewHint from '../../ViewHint.js';
 import {
   apply as applyTransform,
   compose as composeTransform,
-  makeInverse,
-  toString as toTransformString,
 } from '../../transform.js';
 import {
   containsCoordinate,
@@ -167,39 +165,13 @@ class CanvasImageLayerRenderer extends CanvasLayerRenderer {
     const scaleY =
       (pixelRatio * imageResolutionY) / (viewResolution * imagePixelRatio);
 
-    const extent = frameState.extent;
-    const resolution = viewState.resolution;
-    const rotation = viewState.rotation;
+    this.prepareContainer(frameState, target);
+
     // desired dimensions of the canvas in pixels
-    const width = Math.round((getWidth(extent) / resolution) * pixelRatio);
-    const height = Math.round((getHeight(extent) / resolution) * pixelRatio);
-
-    // set forward and inverse pixel transforms
-    composeTransform(
-      this.pixelTransform,
-      frameState.size[0] / 2,
-      frameState.size[1] / 2,
-      1 / pixelRatio,
-      1 / pixelRatio,
-      rotation,
-      -width / 2,
-      -height / 2,
-    );
-    makeInverse(this.inversePixelTransform, this.pixelTransform);
-
-    const canvasTransform = toTransformString(this.pixelTransform);
-
-    this.useContainer(target, canvasTransform, this.getBackground(frameState));
+    const width = this.context.canvas.width;
+    const height = this.context.canvas.height;
 
     const context = this.getRenderContext(frameState);
-    const canvas = this.context.canvas;
-
-    if (canvas.width != width || canvas.height != height) {
-      canvas.width = width;
-      canvas.height = height;
-    } else if (!this.containerReused) {
-      context.clearRect(0, 0, width, height);
-    }
 
     // clipped rendering if layer extent is set
     let clipped = false;
@@ -258,10 +230,6 @@ class CanvasImageLayerRenderer extends CanvasLayerRenderer {
       context.restore();
     }
     context.imageSmoothingEnabled = true;
-
-    if (canvasTransform !== canvas.style.transform) {
-      canvas.style.transform = canvasTransform;
-    }
 
     return this.container;
   }
